Restructure sources
