+++ build.gradle
@@ -4,21 +4,14 @@ plugins {
 }
 
 java {
-	toolchain {
-		languageVersion = JavaLanguageVersion.of(8)
-	}
-}
-
-sourceSets {
-	main.java.srcDirs = ['src']
-	main.resources.srcDirs = ['src']
+	toolchain.languageVersion = JavaLanguageVersion.of(8)
 }
 
 application {
-	mainClass  = 'breakout.Main'
+	mainClass = 'breakout.Main'
 }
 
 dependencies {
-	implementation files('acm.jar')
+	implementation fileTree('libs')
 }
