+++ src/main/java/breakout/bricks/Brick.java
