+++ src/main/java/breakout/geometry/Label.java
