+++ src/main/java/breakout/geometry/Point.java
