+++ src/main/resources/breakout/pictures/BronzeBlock.png
