+++ src/main/resources/breakout/pictures/GoldBlock.png
