+++ src/main/resources/breakout/pictures/LaserBall.png
