+++ src/main/resources/breakout/pictures/LightOff.png
