+++ src/main/resources/breakout/pictures/LightOn.png
