+++ src/main/resources/breakout/pictures/MetalBlock.png
