+++ src/main/resources/breakout/pictures/NeonBlock.png
