+++ src/main/resources/breakout/pictures/SilverBlock.png
