+++ src/main/resources/breakout/pictures/paddle.png
